Fix CR issues
