+++ src/Serializer/PayloadSerializerInterface.php
@@ -9,8 +9,10 @@ use Sentry\Event;
 /**
  * This interface defines the contract for the classes willing to serialize an
  * event object to a format suitable for sending over the wire to Sentry.
+ *
+ * @internal
  */
-interface EventSerializerInterface
+interface PayloadSerializerInterface
 {
     /**
      * Serializes the given event object into a string.
