+++ src/Tracing/Transaction.php
@@ -50,54 +50,6 @@ final class Transaction extends Span
     }
 
     /**
-     * @return array<string, mixed> Returns the trace context
-     *
-     * @psalm-return array{
-     *     data?: array<string, mixed>,
-     *     description?: string,
-     *     op?: string,
-     *     parent_span_id?: string,
-     *     span_id: string,
-     *     status?: string,
-     *     tags?: array<string, string>,
-     *     trace_id: string
-     * }
-     */
-    public function getTraceContext(): array
-    {
-        $result = [
-            'span_id' => (string) $this->spanId,
-            'trace_id' => (string) $this->traceId,
-        ];
-
-        if (null !== $this->parentSpanId) {
-            $result['parent_span_id'] = (string) $this->parentSpanId;
-        }
-
-        if (null !== $this->description) {
-            $result['description'] = $this->description;
-        }
-
-        if (null !== $this->op) {
-            $result['op'] = $this->op;
-        }
-
-        if (null !== $this->status) {
-            $result['status'] = $this->status;
-        }
-
-        if (!empty($this->data)) {
-            $result['data'] = $this->data;
-        }
-
-        if (!empty($this->tags)) {
-            $result['tags'] = $this->tags;
-        }
-
-        return $result;
-    }
-
-    /**
      * Attaches SpanRecorder to the transaction itself.
      */
     public function initSpanRecorder(): void
